Add files via upload
The experiment process is designed to explore the dynamics of coalitions formation in fractional hedonic games. 
By systematically varying parameters and tracking outcomes, the experimental study aims to understand under which conditions stable outcomes emerge, how frequently they emerge and which is the number of deviations needed to reach these stable outcomes. 
The insights gained from these experiments help in understanding strategic behavior in the considered hedonic games and confirm the effectiveness of the introduced notions of stability for recovering the existence of and improving the speed of convergence to stable outcomes.

The experiment process starts by randomly generating a weighted (or unweighted) graph with $n$ nodes and whose edges are established based on a given probability $p$ (an edge between any pair of nodes is created with probability $p$).  
This setup allows for the creation of graphs with different values of edge density, depending on the value of $p$. 
Then an initial partition of agents into coalitions (groups) is defined. 
The experiment proceeds with a random sequence of profitable deviations starting from the initial partition. 
The algorithm iteratively checks if there exists a group of agents that can perform a profitable deviation by going through a uniformly sampled sequence of subsets of agents. This random sampling ensures a thorough exploration of possible coalition formations when executing multiple dynamics over the same instance. 
During the dynamics, the algorithm checks for cycles by keeping track of previously encountered outcomes.
The process continues until no further improvements can be made (in this case, the total number of deviations performed is reported) or a cycle is detected. \\ 


{\noindent \bf Parameters.} 
The experimental work explores a variety of parameters. 
The key parameters considered are: $n$ (number of agents), $p$ (edge probability), the edge weights, the initial coalition structure and the type of deviations. 

The number of agents $n$ takes values in the set $\{5, 10, 15, 20, 30, 40, 50, 60,$ $70, 80, 90, 100\}$ (for in $k$-improvement core dynamics, in which the cardinality of the set of possible blocking coalition is exponential in $n$, we limit the number of agents at $20$). This range allows the study to cover small to relatively large group scenarios, assessing how group size impacts coalition dynamics. 
The probability $p$ of an edge existing between any two agents is considered in the set $\{0.05,0.1,0.2,0.3,0.5\}$. This parameter determines the density of the graph, influencing the likelihood and number of connections between agents.
The experiment distinguishes between weighted and unweighted graphs. In weighted graphs, edges have random weights selected from the range $[1,100]$, while in unweighted graphs, all edges are treated equally without specific weights.
Two types of initial coalition structures are considered:
\emph{singleton coalitions} and \emph{maximal connected components}.
In the first case, each agent starts in their own coalition, i.e., every agent is a singleton coalition: this setup simulates a scenario where agents do not have any predefined group affiliations.
In the second case, agents are grouped into coalitions based on the maximal connected components of the underlying graph. In this case, each coalition consists of all agents who are directly or indirectly connected by a path of edges in the graph. Notice that, when the graph is connected, the initial coalition structure contains only the grand coalition.  
The choice between these two initial states provides insights into how starting conditions affect the evolution of the dynamics. 
The experiment examines two types of deviations:
$q$-size core deviations, for $q$ in $\{2,3\}$ and  
$k$-improvement core deviations for $k$ in $\{1, 3/2, 2\}$ (notice that $1$-improvement core deviations correspond to classical core deviations).
For every possible combination of the parameters we sample $50$ instances and run $50$ dynamics and we outcome the average and the standard deviation of the number of deviations performed.